+++ alf/environments/suite_dmlab.py
@@ -17,8 +17,7 @@ import gym
 import gin.tf
 from tf_agents.environments import suite_gym
 from tf_agents.environments import wrappers
-from alf.environments.suite_socialbot import ProcessPyEnvironment
-from alf.environments.suite_socialbot import UnwrappedEnvChecker
+from alf.environments.utils import UnwrappedEnvChecker, ProcessPyEnvironment
 
 _unwrapped_env_checker_ = UnwrappedEnvChecker()
 
@@ -209,7 +208,6 @@ def load(scene,
     Returns:
         A PyEnvironmentBase instance.
     """
-    global _unwrapped_env_checker_
     _unwrapped_env_checker_.check_and_update(wrap_with_process)
 
     if max_episode_steps is None:
