+++ alf/environments/suite_mario.py
@@ -18,11 +18,10 @@ import gin
 
 from tf_agents.environments import suite_gym
 from tf_agents.environments import wrappers
-from alf.environments.suite_socialbot import ProcessPyEnvironment
 from alf.environments.mario_wrappers import MarioXReward, \
     LimitedDiscreteActions, ProcessFrame84, FrameFormat
 from alf.environments.wrappers import FrameSkip, FrameStack
-from alf.environments.suite_socialbot import UnwrappedEnvChecker
+from alf.environments.utils import UnwrappedEnvChecker, ProcessPyEnvironment
 
 _unwrapped_env_checker_ = UnwrappedEnvChecker()
 
@@ -82,7 +81,6 @@ def load(game,
     Returns:
         A PyEnvironmentBase instance.
     """
-    global _unwrapped_env_checker_
     _unwrapped_env_checker_.check_and_update(wrap_with_process)
 
     if spec_dtype_map is None:
