+++ alf/environments/suite_socialbot.py
@@ -19,136 +19,22 @@ try:
 except ImportError:
     social_bot = None
 
-import sys
-import traceback
 import contextlib
 import socket
 import gym
 from fasteners.process_lock import InterProcessLock
-from tf_agents.environments import suite_gym, wrappers, parallel_py_environment
+from tf_agents.environments import suite_gym, wrappers
 import gin.tf
-import tensorflow as tf
-from absl import logging
+
+from alf.environments.utils import UnwrappedEnvChecker, ProcessPyEnvironment
 
 DEFAULT_SOCIALBOT_PORT = 11345
-
-
-class UnwrappedEnvChecker(object):
-    """
-    A class for checking if there is already an unwrapped env in the current
-    process. For some games, if the check is True, then we should stop creating
-    more envs (multiple envs cannot coexist in a process).
-
-    See suite_socialbot.py for an example usage of this class.
-    """
-
-    def __init__(self):
-        self._unwrapped_env_in_process = False
-
-    def check(self):
-        assert not self._unwrapped_env_in_process, \
-            "You cannot create more envs once there has been an env in the main process!"
-
-    def update(self, wrap_with_process):
-        """
-        Update the flag.
-
-        Args:
-            wrap_with_process (bool): if False, an env is being created without
-                being wrapped by a subprocess.
-        """
-        self._unwrapped_env_in_process |= not wrap_with_process
-
-    def check_and_update(self, wrap_with_process):
-        """
-        Combine self.check() and self.update()
-        """
-        self.check()
-        self.update(wrap_with_process)
-
 
 _unwrapped_env_checker_ = UnwrappedEnvChecker()
 
 
 def is_available():
     return social_bot is not None
-
-
-class ProcessPyEnvironment(parallel_py_environment.ProcessPyEnvironment):
-    """tf_agents ProcessPyEnvironment with render()."""
-
-    def __init__(self, env_constructor, flatten=False):
-        super(ProcessPyEnvironment, self).__init__(
-            env_constructor, flatten=flatten)
-
-    def _worker(self, conn, env_constructor, flatten=False):
-        """It's a little different with `super()._worker`, it closes environment when
-        receives _CLOSE.
-
-        Args:
-            conn (Pipe): Connection for communication to the main process.
-            env_constructor (Callable): env_constructor for the OpenAI Gym environment.
-            flatten (bool): whether to assume flattened actions and time_steps
-                during communication to avoid overhead.
-
-        Raises:
-            KeyError: When receiving a message of unknown type.
-        """
-        try:
-            env = env_constructor()
-            action_spec = env.action_spec()
-            conn.send(self._READY)  # Ready.
-            while True:
-                try:
-                    # Only block for short times to have keyboard exceptions be raised.
-                    if not conn.poll(0.1):
-                        continue
-                    message, payload = conn.recv()
-                except (EOFError, KeyboardInterrupt):
-                    break
-                if message == self._ACCESS:
-                    name = payload
-                    result = getattr(env, name)
-                    conn.send((self._RESULT, result))
-                    continue
-                if message == self._CALL:
-                    name, args, kwargs = payload
-                    if flatten and name == 'step':
-                        args = [tf.nest.pack_sequence_as(action_spec, args[0])]
-                    result = getattr(env, name)(*args, **kwargs)
-                    if flatten and name in ['step', 'reset']:
-                        result = tf.nest.flatten(result)
-                    conn.send((self._RESULT, result))
-                    continue
-                if message == self._CLOSE:
-                    assert payload is None
-                    env.close()
-                    break
-                raise KeyError(
-                    'Received message of unknown type {}'.format(message))
-        except Exception:  # pylint: disable=broad-except
-            etype, evalue, tb = sys.exc_info()
-            stacktrace = ''.join(traceback.format_exception(etype, evalue, tb))
-            message = 'Error in environment process: {}'.format(stacktrace)
-            logging.error(message)
-            conn.send((self._EXCEPTION, stacktrace))
-        finally:
-            conn.close()
-
-    def render(self, mode='human'):
-        """Render the environment.
-
-        Args:
-            mode: One of ['rgb_array', 'human']. Renders to an numpy array, or brings
-                up a window where the environment can be visualized.
-        Returns:
-            An ndarray of shape [width, height, 3] denoting an RGB image if mode is
-            `rgb_array`. Otherwise return nothing and render directly to a display
-            window.
-        Raises:
-            NotImplementedError: If the environment does not support rendering.
-        """
-        return self.call('render', mode)()
 
 
 @gin.configurable
@@ -186,7 +72,6 @@ def load(environment_name,
     Returns:
         A PyEnvironmentBase instance.
     """
-    global _unwrapped_env_checker_
     _unwrapped_env_checker_.check_and_update(wrap_with_process)
 
     gym_spec = gym.spec(environment_name)
